ci: generation of shards and dynamic tasks for running sharded tests
