+++ .gitignore
@@ -99,6 +99,7 @@ app/src/main/assets/dependencies_version.json
 /intellij.gdsl
 
 app/version.txt
+shards.json
 # Editor temporary files
 *~
 \#*#
